logout alert + fix navbar

+++ frontend/src/components/Navbar/UserNavBar/ClientDropDownMenu.jsx
@@ -3,8 +3,11 @@ import styles from './Navbar.css';
 
 const ClientDropDownProfile = () => {
     const changeStatusToNull = () => {
-        window.localStorage.setItem('status', 'null');
-        window.localStorage.removeItem('email');
+        let answer = window.confirm("are you sure you want to logout")
+        if (answer) {
+            window.localStorage.setItem('status', 'null');
+            window.localStorage.removeItem('email');
+        }
     }
     return (
         <div className='flex flex-col dropDownProfile'>
@@ -12,7 +15,7 @@ const ClientDropDownProfile = () => {
 
 
                 {/* To add an element to the DropDown menu */}
-                <li><a className='Profile' href='/Profile'>Profile</a></li>
+                <li><a className='Profile' href='/userprofile'>Profile</a></li>
                 <li><a className='MyCard' href='/MyCard'>My Card</a></li>
                 <li><a className='Logout' href='/' onClick={changeStatusToNull}>Log out</a></li>
             </ul>
